flask_template/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
Flask-Login
Authlib
Werkzeug
PyJWT
